refactor(logs): dedupe log viewer
